add some change and Tests
