+++ src/main/java/game/Die.java
@@ -24,7 +24,6 @@ public class Die {
 
     private int random(int min, int max){
 
-
         Random random = new Random();
         return random.nextInt((max-min)+1)+ min;
     }
