--- src/main/java/game/MGame.java
@@ -1,4 +1,0 @@
-package game;
-
-public class MGame {
-}
